% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/attribution.r
\name{attribution}
\alias{attribution}
\title{Add an attribution label to a plot}
\usage{
attribution(text = "Lincoln Mullen <http://lincolnmullen.com>", size = 0.75,
  color = "gray10")
}
\arguments{
\item{text}{label text}

\item{size}{size of the label text}

\item{color}{color of the label text}
}
\description{
This function adds an attribution label to a plot. You can customize the
text and appearance of the label. This function is adapted from Kieran
Healy's code.
}
\examples{
plot(faithful)
attribution()
}

