% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard}
\alias{clipboard}
\title{Copy R objects to the clipboard}
\usage{
clipboard(x)
}
\arguments{
\item{x}{The object to be written}
}
\description{
Copy R objects to the clipboard
}
\examples{
clipboard("Hello, clipboard")
clipboard(mtcars)
}

