% Generated by roxygen2 (4.0.2): do not edit by hand
\name{course_grades}
\alias{course_grades}
\title{Calculate course grades}
\usage{
course_grades(grades, weights)
}
\arguments{
\item{grades}{A data frame of assignment grades. The first column should be
the students' names. Letter grades should be entered in columns named after
the assignment. The data frame may contain other columns that will not be
used the in the calculation of grades. Strings must not be factors.}

\item{weights}{A named vector of the weights for each assignment. The names
  should be identical to the names of the columns used for assignments in the
  \code{grades} data frame. The weights must sum to \code{1}.}
}
\value{
A data frame of names, numeric, and letter grades.
}
\description{
Given a data frame of assignments and a vector of assignment weights,
calculate course grades using a four-point grading scale.
}
\examples{
assignments <-  data.frame(names = c("Anna", "Bill", "Cara", "Dan"),
                           essay = c("B", "A", "D", "C"),
                           paper = c("A-", "B+", "C", "A"),
                           exam  = c("B-", "C+", "D", "A-"),
                           email = c("a@uni.edu", "b@uni.edu",
                            "c@uni.edu", "d@uni.edu"),
                            stringsAsFactors = FALSE)
weights <- c("essay" = 0.2, "paper" = 0.5, "exam" = 0.3)
course_grades(assignments, weights)
}
\seealso{
\code{\link{grades_4pt}}
}

