% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/course-title-length-ok.R
\name{course_title_length_ok}
\alias{course_title_length_ok}
\title{Check that a course title is under the maximum length}
\usage{
course_title_length_ok(title)
}
\arguments{
\item{title}{The character vector for the title}
}
\value{
TRUE for a course within the maximum length; FALSE otherwise.
}
\description{
George Mason University only allows course names with a maximum of thirty
characters.
}
\examples{
course_title_length_ok("Church and State in America")
}

