% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delay_by.R
\name{delay_by}
\alias{delay_by}
\title{Delay the execution of a function}
\usage{
delay_by(delay = 1, f)
}
\arguments{
\item{delay}{The delay in seconds}

\item{f}{The function to be called}
}
\value{
A delayed version of function \code{f}.
}
\description{
Delay the execution of a function
}
\examples{
f <- delay_by(1, sqrt)
sapply(1:3, f)
}
\references{
This function is borrowed from Hadley Wickham, \emph{Advanced R}
(Chapman and Hall, 2014),
\href{http://adv-r.had.co.nz/Function-operators.html}{section 12.1}.
}

