% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display-code-blocks.R
\name{display_code_blocks}
\alias{display_code_blocks}
\title{Set whether or not to display knitr code blocks and output in a document}
\usage{
display_code_blocks(display = TRUE)
}
\arguments{
\item{display}{TRUE to display the code blocks; false to hide them}
}
\description{
The value passed to this function determines whether or not to display code
blocks, their warnings, and their results in a Knitr document.
}

