% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dot_every.R
\name{dot_every}
\alias{dot_every}
\title{Print a dot during repeated executions of a function}
\usage{
dot_every(n = 10, f)
}
\arguments{
\item{n}{Number of iterations between dots}

\item{f}{The function to be called}
}
\value{
A version of function \code{f} that prints dots.
}
\description{
Print a dot during repeated executions of a function
}
\examples{
f <- dot_every(5, sqrt)
r <- lapply(1:100, f)
}
\references{
This function is borrowed from Hadley Wickham, \emph{Advanced R}
(Chapman and Hall, 2014),
\href{http://adv-r.had.co.nz/Function-operators.html}{section 12.1}.
}

