% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fig_save.r
\name{fig_save}
\alias{fig_save}
\title{Save a ggplot to multiple formats}
\usage{
fig_save(plot, filename = "plot", formats = c("svg", "pdf", "png"),
  width = 11, height = 8.5)
}
\arguments{
\item{plot}{The ggplot object to save}

\item{filename}{The file name without the extension}

\item{formats}{A character vector of formats to save}

\item{width}{The width in the same units that \code{ggsave} uses.}

\item{height}{The height in the same units that \code{ggsave} uses.}
}
\description{
This function saves a ggplot object to multiple formats in the same
dimensions with similar filenames. By default it saves to SVG, PDF,
and PNG.
}
\examples{
require(ggplot2)
plot <- qplot(mpg, wt, data = mtcars)
fig_save(plot)
}

