% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gg_attribution}
\alias{gg_attribution}
\title{Add an attribution to a plot that can be saved with ggsave()}
\usage{
gg_attribution(plot, text = "Lincoln Mullen <http://lincolnmullen.com>",
  size = 0.75, color = "gray10")
}
\arguments{
  \item{plot}{The ggplot object to which the attribution
  should be added}

  \item{text}{Label text}

  \item{size}{Size of the label text}

  \item{color}{Color of the label text}
}
\description{
This function adds an attribution label to a plot. You can customize the
text and appearance of the label.
}
\examples{
require(ggplot2)
plot <- qplot(mpg, wt, data = mtcars)
plot_attributed <- gg_attribution(plot, "Your attribution")
print(plot_attributed)
}

