% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grade_factor.R
\name{grade_factor}
\alias{grade_factor}
\title{Convert grades to a factor with ordered levels}
\usage{
grade_factor(x)
}
\arguments{
\item{x}{A character vector of grades, e.g., \code{c("A+", "F", NA)}}
}
\value{
An ordered factor
}
\description{
Convert a character vector of grades to an ordered factor. This function
checks that the grade values are appropriate. It does accept \code{NA}.
}
\examples{
g <- c("A-", "F", NA, "C+", "B+")
grade_factor(g)
}

