% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grades_4pt.R
\docType{data}
\name{grades_4pt}
\alias{grades_4pt}
\title{Four-point grading scale}
\usage{
data(grades_4pt)
}
\description{
Values for converting between letter grades and a four-point numeric scale.
Use \code{average} when converting from letter grades to numeric values, and
\code{lowest} when converting from numeric values to letter grades.
}
\seealso{
\code{\link{grades_percent}}
}
\keyword{datasets}

