% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grades_percent.R
\docType{data}
\name{grades_percent}
\alias{grades_percent}
\title{Percentage-based grading scale}
\usage{
data(grades_percent)
}
\description{
Values for converting between letter grades and a percentage-based numeric
scale. Use \code{average} when converting from letter grades to numeric
values, and \code{lowest} when converting from numeric values to letter
grades.
}
\seealso{
\code{\link{lookup}}, \code{\link{grades_4pt}}
}
\keyword{datasets}

