% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookup}
\alias{lookup}
\title{Lookup a key in a table and return a value}
\usage{
lookup(x, lookup_table, key = 1, value = 2)
}
\arguments{
\item{x}{The keys to lookup}

\item{lookup_table}{The table of keys and values}

\item{key}{The index or name of the column with the keys}

\item{value}{The index or name of the column with the values}
}
\description{
A common problem is to look up some key and return a value for it. This
function lets you provide a vector of keys to lookup and a table (usually a
data frame) with keys and values. Optionally, it lets you specify which
columns of the table contain the keys and the values.
}
\examples{
my_table <- data.frame(letters = letters[1:5],
                       colors = c("red", "green", "blue", "pink", "orange"),
                       stringsAsFactors = FALSE)
lookup(c("a", "e", "c"), my_table)
lookup(c("green", "blue"), my_table, key = "colors", value = "letters")
}

