% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse-nhgis-codebook.R
\name{parse_nhgis_codebook}
\alias{parse_nhgis_codebook}
\title{Parse an NHGIS codebook}
\usage{
parse_nhgis_codebook(filename)
}
\arguments{
\item{filename}{The path to an NHGIS codebook.}
}
\value{
A data frame with two columns \code{code} and \code{meaning}.
}
\description{
\href{http://nhgis.org}{NHGIS} provides historical census data with fields
named, for example AA001. For each csv file it provides a plain text codebook
which explains the meanings of these codes. This function parses those
codebooks and returns a data frame which can be merged back into the dataset
to get the meanings of the codes.
}
\examples{
sample_codebook <- system.file("extdata",
                               "nhgis0024_ds17_1870_county_codebook.txt",
                               package = "mullenMisc")
parse_nhgis_codebook(sample_codebook)
}

