% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_year.R
\name{round_year}
\alias{round_year}
\title{Round years to decades or centuries}
\usage{
round_year(year, round_to = c("decade", "century"))
}
\arguments{
\item{year}{Year to test (as a number)}

\item{round_to}{Round the year to either the \code{decade} or \code{century}.}
}
\value{
The rounded year as an integer.
}
\description{
This function takes a year as a numeric value and rounds it to the nearest
decade or century.
}
\examples{
set_of_years <- seq(1892, 1923, 4)
round_year(set_of_years) # round to decade by default
round_year(set_of_years, round_to = "century")
}

