% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookup}
\alias{lookup}
\title{Lookup a key in a table and return a value}
\usage{
lookup(x, lookup_table, key = 1, value = 2)
}
\arguments{
\item{x}{The keys to lookup}

\item{lookup_table}{The table of keys and values, either as a data frame with
a key column and a value column, or a named vector.}

\item{key}{The index or name of the column with the keys}

\item{value}{The index or name of the column with the values}
}
\description{
A common problem is to look up some key and return a value for it. This
function lets you provide a vector of keys to lookup and a table (either a
data frame or a named vector) with keys and values. If the lookup table is a
data frame, you can specify which columns of the table contain the keys and
the values: the default is to look for the keys in column 1 and the values in
column 2. If the lookup table is a named vector, then the names are the keys
and the vector values are the lookup values.
}
\examples{
my_table <- data.frame(letters = letters[1:5],
                       colors = c("red", "green", "blue", "pink", "orange"),
                       stringsAsFactors = FALSE)
lookup(c("a", "e", "c"), my_table)
lookup(c("green", "blue"), my_table, key = "colors", value = "letters")

my_vector <- c("a" = "red", "b" = "green", "c" = "blue", "d" = "pink")
lookup(c("a", "d", "c"), my_vector)
}

